requirements.txt
matplotlib
pandas
scikit-learn
plotly
seaborn
numpy
umap-learn
karateclub
torch
torch-geometric
networkx